Split main into modules
